requirements.txt
fastapi[all]==0.115.5
asyncpg==0.30.0
SQLAlchemy==2.0.36
aioredis
jwt
passlib
